Fix the bug of quering last week's data when choose 'Last Monday and 7 days ago'.
Change 'Last Monday and 7 days ago' to 'This Monday and 7 days ago'
and 'Last Monday and 14 days ago' to 'This Monday and 14 days ago'
in the since section in time filter.

Add the config of DOWParseStyle in the utils.py

Noted: parsedatetime expects to have a quantity in front of its units.