Fix the bug of deleting row 'all' and column 'all' in pivot table.
Instead of directly deleting rows and columns in default,
we added some configuration in the column configuration and
row configuration for pivot table.

Users can choose "hide the column 'all'" in column configuration
to hide the columns which contains 'all' in 'th' element.

Users can input the string to indicates which row they wants to
hide in row configuration.

The following files have been modified:
- ColumnControl.jsx: Add the hideOrNot variable to implement hide function
- RowControl.jsx: Add the hide array to implement hide function
- pivot_table.js: Add the related part to parse the parameter and hide
  column and rows