Implement table and pivot table sort.
The original table and pivot sort only exists in frontend, when user refresh
browser, the state of the table is lost, thus we can not take screenshot
as we like.

We need to store table state in server and make it work across browsers.

jQuery DataTable support `stateLoadCallback` and `stateSaveCallback`, and we
do some bookkeeping work in these callbacks. By default, the slice state
is stored in slice `form_data` field, and it will be written to
`Slice.params` database table.

Ref:
- http://refreshmymind.com/datatables-state-save-client-server-side/
- https://datatables.net/reference/option/stateLoadCallback