Implement feature of changing relative position of metrics and columns in pivot table's header.
The user can toggle to choose whether metrics on the top or
columns on the top.

Implement this by changing the level in columns of panda's
dataframe.

The feature of 'hide column all' has been influenced by this new
feature, the solution is doing different operations according to
different conditions.