Fix the bug of colouring properly in row config.
When user apply a colour to a row, the left part of the row
which shows table's group by also changed the colour, which
is not our expectation.

After fixing the bug, the big cube in the left won't change
the other part is changed.

Fixing the bug by changing the way of adding color, previously,
the colour is added to whole `tr`, currently, it is added to
each `th` or `td`.