Fix the bug of unexpected sorting and saving in pivot table with multiple groups.
When user do sorting in pivot table with multiple groups, the
grouped column has a default sorting order which refresh in
every sorting, also caused the sorting can not be saved as user
expected.

Fix the bug by modifying the dataTables.rowsGroup.js, change the
logic in sorting operation and deal with different situations.