Fix pivot table metric wrongly sorted
Previously, to sort the metrics of pivot tables, the following steps
are taken:

1. Our metrics in a pivot table renamed to __index.
2. Table is sorted
3. __index renamed back to original name

However, this resulted in a non-proper order due to String comparison,
e.g. (__1, __10, __2)

A sorted list of uuid is now generated and used to map the metrics
instead, ensuring that there is no column name conflict and that the
metrics are sorted in a correct order.