Fix the bug of making the column color config take priority over the row color configuration.
After fixing the bug of not applying row configuration for
some certain cells, the user found that the row configuration
take the priority over the column configuration sometimes,
which is not expected.

Fix the bug by adding some check before applying column config,
if there is already row config for this cell, then remove it
first.