Fix the bug of showing the row and column config in every viz type.
When the use opens config panel of a barchart or a heatmap
which is not a type of table or pivot table, the row and
conlumn configuration also appear.

Fix the bug by changing viztype.js to make row and column
configuration only show in the table and pivot table.