Add guest user to dashboard.
The basic design for dashboard sharing feature is, we add an additional
relationship between user and dashboard: guests.

The owners of a dashboard have full access to a dashboard, while the
guests of a dashboards have readonly access, which means, the guests can
only read the dashboard and consume the data from slices inside a
dashboard. He can not add or remove slices, adjust slice layout, etc.

This commit just add necessary field to ORM model, you need to do a
`superset db upgrade` to create necessary db tables and constraints.