Dig a hole for webshot service to take screenshot.
It's hard for webshot to login via google accounts since webshot uses
the headless phantomjs as its browser. Thus we have to specify a way for
webshot to identify itself, and grant proper access when the request
comes from webshot.

We need to grant `can_dashboard` and `can_explore_json` access, and need
a monkey patch in `can_access` function to make it grant permissions for
webshot.