Return a list of dashboards in which user are owner or guest.
By default, the list on `DashboardModelViewAsync` returns a list of
dashboards where user has data source access to at least of one slices
in the dashboards.

For the new design, we want to return a list of dashboards in index page
where users are either owner or guest of a dashboard.