Improve dashboard list filter behavior.
By default, when user access the dashboard list page, it will returns a
list of dashboards that users has datasource access to at least one
slice in dashboard.

When user click and open the dashboard but he didn't has all of the
datasource access to all slices, superset will redirect to a page that
reminds user to request necessary data source access to special data
source.

This doesn't make much sense cause this just mixed the permissions of
user, dashboard and slices together, which make things quite
complicated.

A better approach is, dashboard management should focus only on the
relationships between users and dashboards. There're only two cases
here:
- The logined user is a owner of the dashboard, then he has full
read/access access to that dashboard
- The logined user is just a guest of the dashboard, then he has only
readonly access to that dashboard.

In the dashboard management page, we only show a list of dashboards
of which user is an owner; while in index page, we show a list of
dashboards of which user is either an owner or a guest.