Allowed column config to be applied on finer levels in pivot table
Whenever a column is selected under column configuration, a new Select element
will appear that allows user to select sub-columns in the Pivot Table.
Selecting them will spawn additional Select elements will appear to allow
users to choose from sub-columns of a finer granularity.

Upon selecting the columns to their desired granularity, a user can
then apply the desired column configuration for each column
matching the selection.

To enable our new feature to work, the following actions were
taken:

- In our new Select elements, our options are sub-columns that has to be
  dynamically generated and changed but no existing Control works this way.
  - Dynamic properties were passed through mapStateToProps but often do not
    need to be updated. Hence the mapStateToProps only gets triggered once for
    each Control before it gets deleted.
  - A new param called updateOnFetch has been added, where you can pass in a
    function to update the props of the Control everytime an update takes
    place.
- New functions and states were added to ColumnControl to properly track
  the options and values of the new Select
- Revamped how column configuration was applied in the frontend.
  - Previously, there was only one granularity for column configuration,
    allowing us to do a direct match between the 'column' in the Control
    and the actual columns in the table.
  - Instead, logic have been added to find all valid configurations for
    each column. From these possible configurations, the most fine-grained
    configuration is then chosen to be applied to the column.

Hotfix for column names wrongly displayed when not selecting any Columns

A bug occurs in Pivot table when no columns are selecting resulting in only a
single letter of each column being displayed. This is due to the Pandas
dataframe having different index types based on the 'columns' and 'metrics'
input from a user. When there are no 'columns' in pivot table, mapping the
dataframe columns returns a string instead of a tuple. Resulting in us
obtaining a single letter of the column name instead of the column name