Implement the feature of excluding the rows in progress bar.
When using progress bar in the pivot table and table,
users want to exclude some rows instead of calculating
all rows because data in some rows is too big which
makes the comparison meaningless.

Implement this by adding UI in the ColorPickerControl
and adding logic in pivot_table and table.

Because some tables and pivot tables have progress color
saved in the previous structrue, added the logic to change
the `progress bar styling` structrue first.