Fix the bug of sorting properly in some situation.
Currently, after fixing the bug of `pivot table groups sometimes
separate after sorting`, there are still two bugs in sorting about
reloading saved sortings after that patch.

First one is in pivot table with multiple groups, if the table has
saved sorting on only one column, the sorting will be changed and
need user to reset and save. Fixed this by checking if it is the
initialization state to decide later operations.

Second one is in pivot table with one group, the new default sorting
will overwrite the saved sorting, may caused by the change of dom
structure, fixed this by making the default sorting as null, but the
user still need to reset and save.