Return list of dashboards for admin where user can be guest or owner.
In original design, we only return a list of dashboards where users are
owner of in the dashboard admin page, and return a list of dashboards
where users are either owner or guest of.

And to uniform the index page and dashboard admin page, we return the
same list of dashboards to user, in both index page, and dashboard admin
page.

However, users can not edit or delete dashboard if he is a guest but not
an owner of that dashboard.

Todo:
- Need to hide edit/delete button for dashboards where user is guest.
  This is a little complicated to implement cause we need to modify the
  templating library from flask-appbuilder.
- The total count of list of dashboard is wrong cause sqlalchemy's
  `query.count()` and `len(query.all()` returns different result when
  there're duplicate rows produced by SQL cross join. Trying really hard
  to fix it, however, it a little difficult cause this may be a bug of
  flask-appbuilder.

Ref:
- https://stackoverflow.com/questions/32435346/sqlalchemy-different-value-from-lenquery-all-and-query-count
- https://groups.google.com/forum/?fromgroups=#!topic/sqlalchemy/iqkl3PB10kM%5B1-25%5D