Small fix to email snapshot button.
- Change tooltip and description text
- Workaround adjustment of the order of the buttons to fix a button
  blink problem

The button blink problme is really weird, when user logged in as a guest
of a dashboard and the email button will appears at the last of the
button group. And when we mouse over that email export button, the moust
just blinks for ever, don't know why.

This fix is just a workaround.