Added feature to download pivot table as xls file.
Implementation:

This feature is implemented by reading and converting the HTML <table>
into an excel worksheet with the js-xlsx library on the frontend.

The html tag/css classes for the cells in the pivot table is then mapped
to a corresponding style for each Cell object in the worksheet Object.

Using the an alternate version of js-xlsx that provides styling, the worksheet
is then written as generated as a blob to be saved onto the user's computer.

This feature is available for all pivot tables in both the explore view and
dashboard view.

Issues faced:

There are two desired features in js-xlsx that exist in two different forks on
the same javascript library. As such, this commit uses two version of js-xlsx:

- https://github.com/SheetJS/js-xlsx

This version of the js-xlsx library is the official version and provides
greater read/write support for excel sheet. Using the library, we are
able to load the HTML <table> as an excel Worksheet directly without
much manipulation

- https://github.com/protobi/js-xlsx

This version provides styling support which is not available in the official
library.

Other changes

- Added http://fontawesome.io/icon/file-excel-o/ icon to the Download xlsx
  button

- Refactored the SliceCell.jsx so that code is not repeated and you can
  easily add a feature to both Guest and non-guests at the same time