Revert "Disable automatically permission add when add sqla table."
This reverts commit 35d85ec6bba0fa76530b5fb1ebe6cdfeb13e3ab7.

The original approach for permissions about newly created tables is,
when users create a new table, the data source access to that table are
added to all user's roles automatically (except for roles in
`ROBOT_PERMISSION_ROLES`).

This approach is implemented by Song Yan.

Apparently, this approach will cause a trouble when user's with multiple
country roles try to create a new table, the data source access to that
newly created table will be added to all country roles of the user,
which, may cause data access leak.

So there're two approaches:

- We disable Song Yan's approach, that is, we disable the automatic
  permission add to user's roles, so here is the process for adding a
  new table:
  - users add a new table
  - the data source access to that new table are added to some roles by
    country admins.
- We still stick to Song Yan's approach, however, we blacklist our own
  'Guest' and 'Editor' role. Besides, we disable by convention users
  with multiple country roles to add new tables.

A better approach to solve this problem is, I think, users with multiple
country roles should specify the necessary one country role that assigned
the data source access to the newly created table, just like when you
create a new repository in github/gitlab, you should specify which
user/organization that the newly created repo belongs to.