Hotfix for javascript error when exporting table with hidden cells to xls
It is possible for the HTML table in a pivot table slice to have hidden
columns. When this occur, this result in the actual dimensions of the
HTML <table> not matching that of the Pivot Table. Resulting in array out of
bound error when attempting to convert the table into a usable excel format.

A function has been added to remove such hidden columns before passing the
HTML table to the export xls functions.

Also removed odd/even colouring as the colouring doesnt work well with
irregular tables