Fixed superset using local user time to filter Queries in database
After submitting a query, Superset makes an async call:

/superset/queries/<user_current_time>

Superset will then search through the submitted Queries filtered table and
return queries that were last updated AFTER the user current time.

If a user's localtime is faster than Superset's server time. The user will
never be able to obtain the query results as it is possible that the last
BEFORE the 'user current time' that was previously obtained.

This is fixed by adding a new action the handle the Async call. If the async
call succeeds, the execute time for the query is saved instead of using the
current local time.