Fixed datasource permissions not being created properly on SQL Lab
The datasource permission is created as a result of an event listener tagged
to the 'after_insert' event when creating a new data source.

Source: superset/connectors/sqla/models.py:761

This event listener calls the 'set_perm' function which creates the permission
using target.get_perm() as the permission name where the target in a SqlaTable
class.

Source: superset/models/helpers.py:120

However, as 'after_insert' takes place before the actual commit in the
database. The get_perm function returns a `None` for the database property as
the database column is a backreference to the db and will not be populated
until the table is properly committed. This creates an improper perm property
for the SqlaTable as such [None].[<TABLE_NAME>]

This bug is fixed by handling the exception and manually getting the database
name to add to the permission name when the database property is not available.