Added feature to sort overall time series metric
Previously, the `sort by` control for time series is misleading to users; it is
only used to sort and limit the number of time series that appear in the final
chart, with the overall sorting based on the first metric in the time series.

A new column control has been added, allowing us to an overall sorting for each
time series slice, allowing time series slices with different metrics to follow
a uniform order.

As a result of this change, the explore view UI for a time series slices has
also been slightly changed.