Fix sqllab-created tables incompatbility with Column Configuration Control
When we visualize a result from SQL Lab, we are directly brought to the explore
view. From there, an error would arise if the viz_type is converted to Table or
Pivot Table or any other visualisation that makes use of the Column
Configuration Control.

This bug arises when the control attempts to read this.props.columns which is
not available when the explore view is accessed from SQL Lab, resulting in a JS
error causing the page to crash.

The bug is fixed by passing an empty list if this.props.columns is unavailable.