Fix presto-generated visualisation using invalid syntax for date comparison
Previously, whenever a Presto visualization is generated on Lumos, an invalid
query will be generated as Lumos tries to do a date filter by directly using
inequality operators to compare dates. This syntax is wrong in Presto.

To fix this issue, two changes were made:

1) convert_dttm for PrestoEngineSpec has been reverted to follow upstream.

This function was commented out such that all `datetime` columns in Presto
datasources will be directly used as is. Reverting the comments would allow
`timestamp` and `date` columns to be enclosed with the required SQL functions
to properly generate the timestamp/date for proper comparison.

However, this does not solve the issue yet as SQL Lab automatically converts
all `timestamps` and `date` columns into `string` as the retrieved data will be
converted to a Pandas Dataframe before being passed as JSON in the XHR
response. Given that these columns are `Strings`. The enclosing of these
columns with the appropriate `date`/`timestamp` functions will not take place.

2) An additional check is done for temporal columns that are integers, bigints
or doubles.

This check is added as we do use these data types as our temporal columns in
our datasource.

3) Column type is now preserved between SQL Lab and original table.

The column types of the original table will now be preserved when you do a
SQL Lab query using Presto. This would allow the previous problem to be solved.
This change is only limited to Presto engine to ensure maximum compatibility.