Fixed bug that occurs when downloading pivot tables with hidden columns.
A bug has been introduced that causes xls download to fail for Pivot tables
whenever columns are hidden. As a result of the bug, an index out of bound
error occurs as it attempted to format the column that has been removed from
the excel sheet.

A fix has been added to properly count the remaining columns before formatting
them.