Two bugs are fixed in this patch.
- Fix the bug of choosing the correct database when users
  click the link of saved query.
- Fix the bug of choosing the database of last active query
  when adding a new query editor in SqlLab.

Previously, when users click the link of saved query, a query
id will be passed to frontend, but after that, an `autoselect`
setting will be executed to choose the first database in the
list. Similarly, when users add a new query in the sql editor,
the database id of active query editor will be passed to new
query editor, but after that, an `autoselect` setting is
executed and set the database to the first one in the list.

As mentioned above, two bugs ate fixed by disabling `autoselect`.