Removed permission is denied error when logging in successfully
Originally, a flask `flash` message is sent over whenever a user tries to reach
a page that he does not has access to. This `flash` message will be displayed
when he is redirected to the login page.

However, as we have switched our login page to use Google oauth, the login page
is an external page which does not consume the flash message. This results in
the flash message being displayed when the user is redirected back to our
application page. However, through the login process, the first application
page our visitors will see would be the welcome page, after a successful login,
which is misleading to our users.

As our login process is altered such that our users will never see a `Lumos`
page until he has successfully logged in. There is no longer any purpose to the
flash message.