Enabled full table xlsx download even with pagination
The current excel download feature works client-side. It functions by finding
the rendered table in the browser and map each individual cells with its
formatting to XLS using the xlsx.js library.

In paginated tables, not all data is not present in the displayed table. To
enable the download of all rows in the tables, a custom option of showing
1 million rows is injected into the option and then selected before the
download. Once the download is completed, the injected option is deleted and
the original option will be selected.