Fix SQL Lab failing when one of the row types is in `bytes` format
As a result of having to differentiate data types for Presto visualizations to
work, data types from DB is preserved when making a SQL Lab query using Presto.

See commit 140dc522

This, together with an upgrade to Python 3 introduced the possiblility of a
'bytes'-type column generated when you do a SQL Lab Query on Presto. This is
due to the fact that `bytes` is not serializable.

To solve this issue, an if branch has been added to solve the issue by decoding
the `bytes` data to an utf8 `string`.