Changed BaseEngineSpec's `limit method` to WRAP_SQL
Superset uses the property from `BaseEngineSpec`.`limit_method to determine
which method it uses to limit SQL queries done on SQL Lab.

The current default method is `FETCH_MANY` which submits the query untouched
and then limiting the query by selecting the desired columns count using the
cursor.

To reduce the load on our datasources and memory load on Superset, the default
limit_method is now changed to `WRAP_SQL`.

`WRAP_SQL` works by wraps the submitted query using an inner query and limits
the outer query. The function is already provided by Superset and does not
require any development on our side.