Fixed mapbox not working on production
Fix a bug where Mapbox doesnt render tile graphics when it is on production.

A `Uncaught ReferenceError: t is not defined` when we visit a Mapbox slice on
production and only production. This is due to the fact that the additional
plug-in used in production, `webpack.optimize.UglifyJsPlugin` is not compatible
with mapbox-gl, a js library required for Mapbox slice to function.

A configuration tweak was found that can remove the error.

Ref: https://github.com/mapbox/mapbox-gl-js/issues/4359