Add google analytics for lumos site monitoring.
Google analytics is the most powerful, and defac to tools for website
statistics.

However, one main problems for lumos RUM (Real User Monitoring) is that,
big dashboards contains tens of queries, while some of the queries need
several minutes to get response. GA is not able to calculate this kind
of AJAX response time.

I've tried to find whether there is some other tools to monitor pages
with long AJAX requests, however, it seems that there're no easy tools
for this job.

So let's first integrate GA and then make some more investigations on
this.