Fix the bug of downloading dashboard as csv.
It happens because we use 'write' instead of 'writestr' when
write csv file with bytes to zip file. The previous method
is used to write a whole file into zip file.

Notes: if rendering data to a slice has an error such as
'no data', it can not be downloaded successfully.