Implement the feature of making sqllab resizable.
Currently the position and the height of sql editor and
result set is fixed. It is kind of annoying when user
has a big result which consists of a lot of rows.

This feature makes user can resize the height of north
pane and south pane to see more rows in the south pane.

Implemented by adding a new component 'react-split-pane'
and change some component behaviours.

Besides, some UI adjustment related to this feature are
also did in this patch.

Problems and solutions:
- When users drag the split line too fast, there will be
  some rendering delay which makes some component overlap
  or produce a big distance between components.
  Resolved by adding a `ondragfinish` function besides
  `onchange` function.
- When users drag the split line too fast, the ace editor
  spreads too fast as well, which makes there are some
  white space inside it.
  Resolved by upgrading the version of ace editor.
- When users drag the split too low, it will disappear.
  Resolved by adding some height limit.
- Fix some conflicts.
- Fix some syntax error.