Disable dashboard refresh when browser window resizes.
Currently, when resizing browser window, the dashboard will re-render
all slices, which will send quite a few request to lumos backend and
sometimes it's quite slow for big dashboard to re-render all slices due
the the default limit of browser about the number of sending requests at
one time.

Just a simple comment to disable this.