Added feature to redirect to respective url with multiple logins
The bug comes from flask-appbuilder whereby the
AuthOAuthView function ignores all parameters
passed by the user. Replace the login function
to consider the next param.

Used a hack to trick google authentication to
accept the parameter by passing it into state.
There is a problem with state that Google will
reject redirect when the states are different.
Repeat redirection to go to the correct url
from flask-appbuilder after checking if user
is logged in.