Add the feature to let user close tab more easily.
When the user hover the mouse on left part of tab in
sql editor, where there is a solid circle, the close
mark will appear. Users can click this button to close
the tab.

Previously, users can only close the tab in the drop
down menu, which is not obviouse at all.

Implement the feature by adding a new component
TabStatusIcon.