Fixed barchart total value rendering and overlapping problem for multiple lines legend
There are currently 2 problems in barchart.

The first problem is the legend is rendered
after labelling the barchart total. The offset
can only be calculated after finish rendering
or on change.

The second problem is the fd.stacked value is
flawed. If controls are enabled for user to
pick grouped or stacked from front-end, the
stacked will retain the last state before
queried from backend. Replacing the stacked
value to obtain from front-end fixes the
problem.