Fixed bug where non-admin cannot access table after changing table name
The problem comes from when user edits a table, the updates will only
be altered in the tables table in psql. However when the tables are
filtered in the table menu and access control, it is retrieved from
ab_view_menu table. The name in ab_view_menu isn't updated hence the
table can not be found. Fixed it by making sure when the table name
is updated, ab_view_menu is updated as well.

Official superset uses another fix by adding the permission to
ab_view_menu if it doesn't exist which is not a good way because
it will just accumulate the table size. Since lumos has a fix in
making sure people can't use the same table name, I think this is
the best possible solution.