Implement the vertical split line in SqlLab in order to let user see full name of table.
Currently the width of left part and right part is fixed
in sqllab, sometimes when the table name is too long, users
have to scroll it to see full name.

Implement the feature by adding a new split pane in sqllab.

Besides, changed some UI style to make it looks better.