Fix the bug of not showing y axis label in some situations.
Currently if the user choose a really short y axis format,
lumos will calculate a value named `maxYAxisLabelWidth`,
which is used to decide the margin left value, causing the
margin left is too narrow to display y axis label.

Fixed by setting a default maxYAxisLabelWidth when the
calculation value is too small. Choose 25 as default
maxYAxisLabelWidth because the value is used somewhere else
also which is suitable for this case.