Implement the feature of showing details of column type `dictionary` when users do query in sqllab.
Currently, when users query for presto in sqllab, if there
are complex column types in the result, it doesn't show the
detailed structure of it which makes user hard to write query
sentence.

In this patch, users can see the `object` type in the left bar,
and see the details in the table header of the preview section
and result section.